requirements.txt
dashscope>=1.17.0
python-dotenv>=1.0.0
PyYAML>=6.0
Jinja2>=3.0
streamlit>=1.28.0
rich>=13.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0
aiosqlite>=0.19.0,<0.22
chromadb>=0.5.0
